Generated VTK files from DigitalPhantom
They've been added to the corresponding folder. I've checked that
the result is consistent with the volumes in freeview and Slicer4.

Close #102. Now it is a matter of clarifying and assessing the way
ITK is accesing the vertices. Therefore, in #92 I will check the
necessary inputs to make the workflows consistent.