Closed forever orientation issues.
Now it is clear: ITK uses LPS convention, and reverses the
matrix when loading volumes. Conversely, VTK files are assumed
to be in world coordinates directly.

Then, Freesurfer has RAS convention. This means that a freesurfer file
with RAS orientation will be an image with LPS orientation matrix in
ITK.

Additionally, mris_convert does not add the geometrical information
to the vertices. Thus, the generated vtk files are in the so-called
tkRAS coordinate space (the one of freeview and such tools).

A new function named fixvtk has been added to transform all the
vertices to a target scannerRAS coordinates (world space).

I tested it on one case, and regions generated by ACWEReg are consistent
with the resampled volume that is internally used in energy computations.

This enables that we generate the datasets of #118